rename: client 파일명 혼란을 줄이기 위해 수정 (#159)
