+++ apps/web/lib/prisma/index.ts
@@ -1,1 +1,1 @@
-export * from './prisma-client';
+export * from './prisma';
